+++ CHANGELOG.md
@@ -23,6 +23,9 @@
 - Add DB schema change check ([#1268](https://github.com/0xMiden/miden-node/pull/1485)).
 - Improve DB query performance for account queries ([#1496](https://github.com/0xMiden/miden-node/pull/1496).
 - The network monitor now marks the chain as unhealthy if it fails to create new blocks ([#1512](https://github.com/0xMiden/miden-node/pull/1512)).
+- Block producer now detects if it is desync'd from the store's chain tip and aborts ([#1520](https://github.com/0xMiden/miden-node/pull/1520)).
+  - This could occur in rare situations where the block is submitted to the store, and the request is only cancelled from the block-producer's perspective.
+  - For example, due to a gRPC timeout.
 
 ### Changes